pyproject.toml
[tool.poetry]
name = "ytdrama"
version = "0.1.0"
description = ""
authors = ["Br0k3nDesign <adam.j.moran@gmail.com>"]
readme = "README.md"
packages = [{include = "ytdrama"}]


[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
google-api-python-client = "^2.77.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
youtube-transcript-api = "^0.5.0"
python-dotenv = "^0.21.1"
sqlalchemy = "^2.0.3"
python-box = "^7.0.0"
pyfunctional = "^1.4.3"
python-dateutil = "^2.8.2"
nltk = "^3.8.1"
connectorx = "0.3.2a2"
pyarrow = "^11.0.0"
cupy-cuda12x = "^11.5.0"
spacy = { file = "local/spacy-3.5.0-cp311-cp311-win_amd64.whl" }
spacy_legacy = { file = "local/spacy_legacy-3.0.12-py2.py3-none-any.whl"}
spacy_loggers = { file = "local/spacy_loggers-1.0.4-py3-none-any.whl"}
spacy_lookups_data = { file = "local/spacy_lookups_data-1.0.3-py2.py3-none-any.whl"}
wordcloud = { file = "local/wordcloud-1.8.1-cp311-cp311-win_amd64.whl" }

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
